requirements.txt
absl-py==1.0.0
aioredis==1.3.1
amqp==5.1.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
autobahn==22.3.2
Automat==20.2.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
billiard==3.6.4.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.0.0
caffeine==2.9.8
celery==5.2.6
certifi==2020.6.20
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.0
cupshelpers==1.0
cycler==0.11.0
daphne==3.0.2
dbus-python==1.2.18
debugpy==1.6.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.7.0
distro-info===1.1build1
Django==4.0.4
django-celery-results==2.3.1
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
duplicity==0.8.21
entrypoints==0.4
ewmh==0.1.6
executing==0.8.3
fasteners==0.14.1
ffmpeg-python==0.2.0
flake8==4.0.1
flatbuffers==1.12
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
gyp==0.1
h5py==3.6.0
hiredis==2.0.0
html5lib==1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
imageio==2.4.1
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.3.0
itypes==1.2.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==0.17.0
jupyter-client==7.3.0
jupyter-core==4.10.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.3.2
kombu==5.2.4
labelImg==1.8.6
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.1
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mediapipe==0.8.10
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
nest-asyncio==1.5.5
netifaces==0.11.0
nose==1.3.7
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
-e git+https://github.com/onnx/onnx.git@8ab1efa4fc979e1748ebd378a71a59a22deabe88#egg=onnx
opencv-contrib-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
Pivy==0.6.5
pluggy==1.0.0
ply==3.11
prompt-toolkit==3.0.29
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.0
PyJWT==2.4.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyRFC3339==1.1
pytest==7.1.2
pytest-django==4.5.2
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.20.0
python-xlib==0.29
python3-openid==3.2.0
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
redis==4.3.1
repo==2
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==0.23.2
scipy==1.8.0
SecretStorage==3.3.1
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
supervisor==4.2.1
sympy==1.9
systemd-python==234
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
torch==1.11.0+cpu
torchaudio==0.11.0+cpu
torchvision==0.12.0+cpu
tornado==6.1
traitlets==5.1.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
Werkzeug==2.1.1
whitenoise==6.0.0
wrapt==1.14.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
